requirements.txt
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
configparser==3.8.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.5
django-ckeditor==5.7.1
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-js-asset==1.2.2
django-money==0.15
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
openapi-codec==1.3.2
Pillow==6.1.0
py-moneyed==0.8.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
xattr==0.9.6
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bs4==0.0.1
datefinder==0.7.0
DateTime==4.3
mysql==0.0.2
protobuf==3.6.1
python-dateutil==2.7.5
regex==2019.1.24
six==1.12.0
soupsieve==1.7.3
zope.interface==4.6.0

